+++ api/api/api_workspace.py
@@ -16,6 +16,8 @@ from typing import List
 from pydantic import BaseModel
 from db_models.workspace import CurrentWorkspace
 from api.api_user import get_current_user, User as UserModelSerializer
+from db_models.workspace import CurrentWorkspace
+from db_models.deals import Deal
 
 current_workspace_router = APIRouter()
 
@@ -38,6 +40,9 @@ class CurrentWorkspaceResponse(BaseModel):
 
 @current_workspace_router.post("/current_workspace/", response_model=CurrentWorkspaceResponse)
 def add_current_workspace(item: CurrentWorkspaceCreate, db: Session = Depends(get_db),current_user: UserModelSerializer = Depends(get_current_user)):
+    data=db.query(Deal).filter(Deal.id==item.deal_id).first()
+    if str(data.user_id) != current_user.id:
+        raise HTTPException(status_code=404, detail="You are not authorized to add workspace")
     workspace = CurrentWorkspace(**item.dict())
     db.add(workspace)
     db.commit()
@@ -47,6 +52,9 @@ def add_current_workspace(item: CurrentWorkspaceCreate, db: Session = Depends(ge
 @current_workspace_router.put("/current_workspace/{workspace_id}", response_model=CurrentWorkspaceResponse)
 def update_workspace(workspace_id: str, item: BaseTableSchema, db: Session = Depends(get_db), current_user: UserModelSerializer = Depends(get_current_user)):
     workspace = db.query(CurrentWorkspace).filter(CurrentWorkspace.id == workspace_id).first()
+    data=db.query(Deal).filter(Deal.id==workspace.deal_id).first()
+    if str(data.user_id) != current_user.id:
+        raise HTTPException(status_code=404, detail="You are not authorized to modify workspace")
     if not workspace:
         raise HTTPException(status_code=404, detail="data item not found")
     workspace.type = item.type
@@ -59,6 +67,9 @@ def update_workspace(workspace_id: str, item: BaseTableSchema, db: Session = Dep
 @current_workspace_router.delete("/current_workspace/{workspace_id}", response_model=CurrentWorkspaceResponse)
 def delete_todo(workspace_id: str, db: Session = Depends(get_db), current_user: UserModelSerializer = Depends(get_current_user)):
     workspace = db.query(CurrentWorkspace).filter(CurrentWorkspace.id == workspace_id).first()
+    data=db.query(Deal).filter(Deal.id==workspace.deal_id).first()
+    if str(data.user_id) != current_user.id:
+        raise HTTPException(status_code=404, detail="You are not authorized to delete workspace")
     if not workspace:
         raise HTTPException(status_code=404, detail="Current workspace not found")
     db.delete(workspace)
@@ -67,6 +78,9 @@ def delete_todo(workspace_id: str, db: Session = Depends(get_db), current_user: 
 
 @current_workspace_router.get("/current_workspace/", response_model=List[CurrentWorkspaceResponse])
 def current_worspace(deal_id: Optional[UUID] = None, db: Session = Depends(get_db), current_user: UserModelSerializer = Depends(get_current_user)):
+    data=db.query(Deal).filter(Deal.id==deal_id).first()
+    if str(data.user_id) != current_user.id:
+        raise HTTPException(status_code=404, detail="You are not authorized to fetch workspace")
     query = db.query(CurrentWorkspace)
     if deal_id:
         query = query.filter(CurrentWorkspace.deal_id == deal_id)
